% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get data from airtable}
\usage{
get_table(API, base, table)
}
\arguments{
\item{API}{character. API key for the airtable account.}

\item{base}{character. ID of the selected base.}

\item{table}{character. Name of the table in the base.}
}
\value{
Returns a data.frame
}
\description{
\code{get_table} is a function used to prepare data.frame from a selected table
in a specific airtable Base.
}
\details{
This function uses the credentials to call airtable API and then
assembles the resulting data.frame using \code{rawToChar} and \code{jsonlite::fromJSON()}.
}
\examples{
\dontrun{
get_table('APIKEY', 'BASEID', 'TABLENAME')
}
}
\author{
Adela Pavelkova & Michal Kubista
}
